requirements.txt
fastapi
azure-functions
requests
